Add files via upload
This is my first project and was created from scratch using Html, CSS and Java Script. Throughout this journey, the MDN articles provided me the necessary information and understanding. (https://developer.mozilla.org/en-US/)

Description of the project:
The page displays three default buttons and one of them is the winner one. After choosing the winner button, you can start the game again.
Additionally, users are allow to generate the desired number of buttons by submitting an input in a form. The same algorithm as before applies - one button is winner and after finding it, one can restart the game.